refactor: Move compact_candidates_with_memory_budget to share with cold
